+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Experimental changelog. Mostly based on [keepachangelog](https://keepachangelog.
 - Added `.mquality` / `.musicquality` - Set encoding quality. Has 4 presets - Low, Medium, High, Highest. Default is Highest
 - Added `.xprewsreset` which resets all currently set xp level up rewards
 - Added `.purgeuser @User` which will remove the specified from the database completely. Removed settings include: Xp, clubs, waifu, currency, etc...
+- Added `.config xp txt.per_image` and xpFromImage to xp.yml - Change this config to allow xp gain from posting images. Images must be 128x128 or greater in size
 
 ### Changed